+++ App/authPopup.js
@@ -14,11 +14,17 @@ function selectAccount () {
 
     const currentAccounts = myMSALObj.getAllAccounts();
 
-    if (currentAccounts.length === 0) {
+    if (!currentAccounts || currentAccounts.length < 1) {
         return;
     } else if (currentAccounts.length > 1) {
         // Add your account choosing logic here
-        console.log("Multiple accounts detected.");
+        console.log("Multiple accounts detected!");
+        currentAccounts.forEach(acc => console.log(acc));
+
+        // Defaulting to the first account found
+        accountId = currentAccounts[0].homeAccountId;
+        username = currentAccounts[0].username;
+        welcomeUser(username);
     } else if (currentAccounts.length === 1) {
         accountId = currentAccounts[0].homeAccountId;
         username = currentAccounts[0].username;
@@ -30,6 +36,7 @@ function selectAccount () {
 selectAccount();
 
 function handleResponse(response) {
+    console.log(response)
     /**
      * To see the full list of response object properties, visit:
      * https://github.com/AzureAD/microsoft-authentication-library-for-js/blob/dev/lib/msal-browser/docs/request-response-object.md#response
@@ -55,16 +62,6 @@ function signIn() {
         .then(handleResponse)
         .catch(error => {
             console.log(error);
-                
-            // Error handling
-            if (error.errorMessage) {
-                // Check for forgot password error
-                // Learn more about AAD error codes at https://docs.microsoft.com/en-us/azure/active-directory/develop/reference-aadsts-error-codes
-                if (error.errorMessage.indexOf("AADB2C90118") > -1) {
-                    myMSALObj.loginPopup(b2cPolicies.authorities.forgotPassword)
-                        .then(response => handlePolicyChange(response));
-                }
-            }
         });
 }
 
@@ -75,13 +72,16 @@ function signOut() {
      * https://github.com/AzureAD/microsoft-authentication-library-for-js/blob/dev/lib/msal-browser/docs/request-response-object.md#request
      */
 
-    // Choose which account to logout from.
-    
+    // Choose which account to logout from and the behavior after logout completes.
     const logoutRequest = {
-        account: myMSALObj.getAccountByHomeId(accountId)
+        account: myMSALObj.getAccountByHomeId(accountId),
+        postLogoutRedirectUri: msalConfig.auth.redirectUri,
+        redirectMainWindowTo: msalConfig.auth.redirectUri
     };
     
-    myMSALObj.logout(logoutRequest);
+    myMSALObj.logoutPopup(logoutRequest).then(() => {
+        window.location.reload();
+    });
 }
 
 function getTokenPopup(request) {
@@ -103,8 +103,7 @@ function getTokenPopup(request) {
             return response;
         })
         .catch(error => {
-            console.log(error);
-            console.log("silent token acquisition fails. acquiring token using popup");
+            console.log("Silent token acquisition fails. Acquiring token using popup. \n", error);
             if (error instanceof msal.InteractionRequiredAuthError) {
                 // fallback to interaction when silent call fails
                 return myMSALObj.acquireTokenPopup(request)
@@ -136,21 +135,20 @@ function passTokenToApi() {
 
 function editProfile() {
     myMSALObj.loginPopup(b2cPolicies.authorities.editProfile)
-      .then(response => handlePolicyChange(response));
+        .then(handlePolicyChange)
+        .catch(error => {
+            console.log(error);
+        });
 }
 
 function handlePolicyChange(response) {
-    /**
-     * We need to reject id tokens that were not issued with the default sign-in policy.
-     * "acr" claim in the token tells us what policy is used (NOTE: for new policies (v2.0), use "tfp" instead of "acr").
-     * To learn more about B2C tokens, visit https://docs.microsoft.com/en-us/azure/active-directory-b2c/tokens-overview
-     */
-
-    if (response.idTokenClaims['acr'] === b2cPolicies.names.editProfile) {
-        window.alert("Profile has been updated successfully. \nPlease sign-in again.");
-        myMSALObj.logout();
-    } else if (response.idTokenClaims['acr'] === b2cPolicies.names.forgotPassword) {
-        window.alert("Password has been reset successfully. \nPlease sign-in with your new password.");
-        myMSALObj.logout();
-    }
+        /**
+         * We need to reject id tokens that were not issued with the default sign-in policy.
+         * "tfp" claim in the token tells us what policy is used (NOTE: legacy policies may use "acr" instead of "tfp").
+         * To learn more about B2C tokens, visit https://docs.microsoft.com/en-us/azure/active-directory-b2c/tokens-overview
+         */
+         if (response.idTokenClaims['tfp'] === b2cPolicies.names.editProfile) {
+            window.alert("Profile has been updated successfully. \nPlease sign-in again.");
+            myMSALObj.logout();
+        }
 }